server/requirements.txt
# ── Web server ────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# ── Local AI: Object detection ────────────────────────────────────────
ultralytics>=8.2.0        # YOLOv8 – downloads yolov8n.pt on first run
Pillow>=10.0.0

# ── Deep Learning backend (CPU-only) ─────────────────────────────────
# For CPU-only install (smaller download):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
